Move server protocols to the server realm
